+++ app/javascript/RootSiteConventionsAdmin/queries.generated.ts
@@ -28,7 +28,7 @@ export type RootSiteConventionsAdminTableQueryQuery = (
 
 export type ConventionDisplayFieldsFragment = (
   { __typename: 'Convention' }
-  & Pick<Types.Convention, 'id' | 'name' | 'starts_at' | 'ends_at' | 'canceled' | 'timezone_name' | 'domain' | 'site_mode' | 'ticket_mode' | 'show_event_list' | 'show_schedule' | 'email_from'>
+  & Pick<Types.Convention, 'id' | 'name' | 'starts_at' | 'ends_at' | 'canceled' | 'timezone_name' | 'timezone_mode' | 'domain' | 'site_mode' | 'ticket_mode' | 'show_event_list' | 'show_schedule' | 'email_from' | 'hidden' | 'language'>
   & { maximum_event_signups?: Types.Maybe<(
     { __typename: 'ScheduledValue' }
     & { timespans: Array<(
@@ -74,12 +74,15 @@ export const ConventionDisplayFieldsFragmentDoc = gql`
   ends_at
   canceled
   timezone_name
+  timezone_mode
   domain
   site_mode
   ticket_mode
   show_event_list
   show_schedule
   email_from
+  hidden
+  language
   maximum_event_signups {
     timespans {
       start
